add new Class ControllableBoardObject
move Properties and Methods to ControllableBoardObject
