+++ AntWars/AntWars.csproj
@@ -71,6 +71,7 @@
     <Compile Include="Board\Ants\Carry.cs" />
     <Compile Include="Board\Ants\Scout.cs" />
     <Compile Include="Board\BoardObjects.cs" />
+    <Compile Include="Board\ControllableBoardObject.cs" />
     <Compile Include="Board\Coordinates.cs" />
     <Compile Include="ConfigurationPanel.cs">
       <SubType>Form</SubType>
