+++ AntWars/Board/Ants/Ant.cs
@@ -10,22 +10,8 @@ namespace AntWars.Board.Ants
     /// <summary>
     /// Die Ameise.
     /// </summary>
-    public abstract class Ant : BoardObject
+    public abstract class Ant : ControllableBoardObject
     {
-        /// <summary>
-        /// Gibt an, ob die Ameise sich in diesem Tick schon bewegt hat. 
-        /// </summary>
-        public bool TookAction { get; internal set; }
-
-        /// <summary>
-        /// Der Faktor für die Berechnung der Bewegungsreichweite.
-        /// </summary>
-        public int MoveRangeFactor { get; internal set; }
-
-        /// <summary>
-        /// Wie weit die Ameise schon gegangen ist.
-        /// </summary>
-        public int UnitsGone { get; internal set; }
        
         /// <summary>
         /// Das Maximale Inventar der Ant.
@@ -37,148 +23,16 @@ namespace AntWars.Board.Ants
         /// </summary>
         public int Inventory { get; protected set; }
 
-        /// <summary>
-        /// Wie weit die Ameise sehen kann.
-        /// </summary>
-        public int ViewRange { get; protected set; }
-
-        /// <summary>
-        /// Lebenspunkte der Ameise.
-        /// </summary>
-        public int Health { get; protected set; }
-
-        /// <summary>
-        /// Angriffstärke der Ameise.
-        /// </summary>
-        public int AttackPower { get; protected set; }
-
-        /// <summary>
-        /// Wie weit die Ameise gehen kann. 
-        /// </summary>
-        public int MoveRange { get; protected set; }
-
         internal Player Owner { get; private set; }
         internal IAIAnt AI { get; set; }
-        internal Board board;
         private Base Base;
 
         internal Ant(Board board, Player owner, int viewRange, int maxInventory, int moveRangeFactor, int hp)
+            : base(board, viewRange, moveRangeFactor, hp)
         {
-            ViewRange = viewRange;
             MaxInventory = maxInventory;
-            this.board = board;
-            this.UnitsGone = 0;
             Owner = owner;
             Inventory = 0;
-            MoveRangeFactor = moveRangeFactor;
-            MoveRange = MoveRangeFactor * board.Diagonal;
-            Health = hp;
-            TookAction = false;
-        }
-
-        /// <summary>
-        /// Lässt die Ameise nach links bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveLeft()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X - 1, Coords.Y);
-            return move(newCoords);
-        }
-        /// <summary>
-        /// Lässt die Ameise diagonal nach links oben bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveUpperLeft()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X - 1, Coords.Y - 1);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise diagonal nach links unten bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveLowerLeft()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X - 1, Coords.Y + 1);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise nach rechts bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveRight()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X + 1, Coords.Y);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise diagonal nach rechts oben bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveUpperRight()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X + 1, Coords.Y - 1);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise diagonal nach rechts unten bewegen.
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist.</returns>
-        public bool moveLowerRight()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X + 1, Coords.Y + 1);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise nach oben bewegen
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist</returns>
-        public bool moveUp()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X, Coords.Y - 1);
-            return move(newCoords);
-        }
-
-        /// <summary>
-        /// Lässt die Ameise nach unten bewegen
-        /// </summary>
-        /// <returns>true wenn das Bewegen erfolgreich war, false wenn etwas im Weg ist</returns>
-        public bool moveDown()
-        {
-            Coordinates newCoords = new Coordinates(Coords.X, Coords.Y + 1);
-            return move(newCoords);
-        }
-
-        private bool move(Coordinates to)
-        {
-            if (!canMove())
-            {
-                die();
-            }
-            else if (!TookAction && board.BoardObjects.move(this, to))
-            {
-                TookAction = true;
-                UnitsGone++;
-                return true;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Prüfen, ob die Ameise sich noch bewegen kann.
-        /// </summary>
-        /// <returns>True wenn ja, False wenn nicht</returns>
-        public bool canMove()
-        {
-            if (UnitsGone >= MoveRange)
-                return false;
-            return true;
         }
 
         /// <summary>
